fix(opengl-tutorial): rename file + add logs for shaders compilation
Removed the triangles and replaced it by a square.
Added log at compilation of the GLSL code to help debugging and see potential issues later on.
