+++ app/src/main/java/com/github/lookupgroup27/lookup/opengl/Square.kt
@@ -8,48 +8,68 @@ import java.nio.ShortBuffer
 
 class Square : Shape() {
 
-  override val color = floatArrayOf(0.0f, 0.0f, 1.0f, 1.0f)
+  override val color = floatArrayOf(1.0f, 1.0f, 1.0f, 1.0f)
 
   // number of coordinates per vertex in this array
   var squareCoords =
-      floatArrayOf(
-          -0.5f,
-          0.5f,
-          0.0f, // top left
-          -0.5f,
-          -0.5f,
-          0.0f, // bottom left
-          0.5f,
-          -0.5f,
-          0.0f, // bottom right
-          0.5f,
-          0.5f,
-          0.0f // top right
-          )
+    floatArrayOf(
+      -0.25f, -0.25f, 0.0f,
+      -0.25f, 0.25f, 0.0f,
+      0.25f, 0.25f, 0.0f,
+      0.25f, -0.25f, 0.0f
+    )
 
-  private val drawOrder = shortArrayOf(0, 1, 2, 0, 2, 3) // order to draw vertices
+  var colorVertex =
+    floatArrayOf(
+      1.0f, 0.0f, 0.0f,
+      0.0f, 1.0f, 0.0f,
+      0.0f, 0.0f, 1.0f,
+      0.0f, 0.0f, 0.0f
+    )
+
+  // Indices in which openGL will draw each triangle vertex by vertex
+  // e.g. It will first draw the triangle with vertices 0, 3, 5
+  // then 3, 2, 4 and so on
+  private val drawOrder =
+    shortArrayOf(
+      0, 1, 2,
+      0, 2, 3
+    ) // order to draw vertices
+
+  private val bytesForOneFloat = 4
 
   // initialize vertex byte buffer for shape coordinates
   private val vertexBuffer: FloatBuffer =
-      // (# of coordinate values * 4 bytes per float)
-      ByteBuffer.allocateDirect(squareCoords.size * 4).run {
-        order(ByteOrder.nativeOrder())
-        asFloatBuffer().apply {
-          put(squareCoords)
-          position(0)
-        }
+    // (# of coordinate values * 4 bytes per float)
+    ByteBuffer.allocateDirect(squareCoords.size * bytesForOneFloat).run {
+      order(ByteOrder.nativeOrder())
+      asFloatBuffer().apply {
+        put(squareCoords)
+        position(0)
       }
+    }
+  private val colorBuffer: FloatBuffer =
+    // (# of coordinate values * 4 bytes per float)
+    ByteBuffer.allocateDirect(colorVertex.size * bytesForOneFloat).run {
+      order(ByteOrder.nativeOrder())
+      asFloatBuffer().apply {
+        put(colorVertex)
+        position(0)
+      }
+    }
+
+  private val bytesForOneShort = 2
 
   // initialize byte buffer for the draw list
   private val drawListBuffer: ShortBuffer =
-      // (# of coordinate values * 2 bytes per short)
-      ByteBuffer.allocateDirect(drawOrder.size * 2).run {
-        order(ByteOrder.nativeOrder())
-        asShortBuffer().apply {
-          put(drawOrder)
-          position(0)
-        }
+    // (# of coordinate values * 2 bytes per short)
+    ByteBuffer.allocateDirect(drawOrder.size * bytesForOneShort).run {
+      order(ByteOrder.nativeOrder())
+      asShortBuffer().apply {
+        put(drawOrder)
+        position(0)
       }
+    }
 
   private var mProgram: Int
 
@@ -59,17 +79,17 @@ class Square : Shape() {
 
     // create empty OpenGL ES Program
     mProgram =
-        GLES20.glCreateProgram().also {
+      GLES20.glCreateProgram().also {
 
-          // add the vertex shader to program
-          GLES20.glAttachShader(it, vertexShader)
+        // add the vertex shader to program
+        GLES20.glAttachShader(it, vertexShader)
 
-          // add the fragment shader to program
-          GLES20.glAttachShader(it, fragmentShader)
+        // add the fragment shader to program
+        GLES20.glAttachShader(it, fragmentShader)
 
-          // creates OpenGL ES program executables
-          GLES20.glLinkProgram(it)
-        }
+        // creates OpenGL ES program executables
+        GLES20.glLinkProgram(it)
+      }
   }
 
   private var positionHandle: Int = 0
@@ -85,27 +105,31 @@ class Square : Shape() {
     vPMatrixHandle = GLES20.glGetUniformLocation(mProgram, "uMVPMatrix")
 
     // get handle to vertex shader's vPosition member
-    positionHandle =
-        GLES20.glGetAttribLocation(mProgram, "vPosition").also {
-          GLES20.glEnableVertexAttribArray(it)
+    positionHandle = GLES20.glGetAttribLocation(mProgram, "vPosition")
+    GLES20.glEnableVertexAttribArray(positionHandle)
 
-          GLES20.glVertexAttribPointer(
-              it, COORDS_PER_VERTEX, GLES20.GL_FLOAT, false, vertexStride, vertexBuffer)
+    GLES20.glVertexAttribPointer(
+      positionHandle, COORDS_PER_VERTEX, GLES20.GL_FLOAT, false, vertexStride, vertexBuffer
+    )
 
-          // get handle to fragment shader's vColor member
-          mColorHandle =
-              GLES20.glGetUniformLocation(mProgram, "vColor").also {
-                GLES20.glUniform4fv(it, 1, color, 0)
-              }
+    // get handle to fragment shader's vColor member
+    mColorHandle = GLES20.glGetAttribLocation(mProgram, "vColor")
+    GLES20.glEnableVertexAttribArray(mColorHandle)
+    GLES20.glVertexAttribPointer(
+      mColorHandle, 3, GLES20.GL_FLOAT, false, 0, colorBuffer
+    )
 
-          GLES20.glUniformMatrix4fv(vPMatrixHandle, 1, false, mvpMatrix, 0)
 
-          // Draw the square
-          GLES20.glDrawElements(
-              GLES20.GL_TRIANGLES, drawOrder.size, GLES20.GL_UNSIGNED_SHORT, drawListBuffer)
+    GLES20.glUniformMatrix4fv(vPMatrixHandle, 1, false, mvpMatrix, 0)
 
-          // Disable vertex array
-          GLES20.glDisableVertexAttribArray(it)
-        }
+    // This time we use draw elements cause it's a composition of multiple triangles
+    // Draw every triangles
+    GLES20.glDrawElements(
+      GLES20.GL_TRIANGLES, drawOrder.size, GLES20.GL_UNSIGNED_SHORT, drawListBuffer
+    )
+
+    // Disable vertex array
+    GLES20.glDisableVertexAttribArray(positionHandle)
+    GLES20.glDisableVertexAttribArray(mColorHandle)
   }
 }
